+++ module/localisation/BallLocalisation/src/BallModel.hpp
@@ -41,7 +41,7 @@ namespace module::localisation {
         static constexpr uint PX = 0;
         static constexpr uint PY = 1;
 
-        static constexpr size_t size = 2;
+        static constexpr int size = 2;
 
         using StateVec = Eigen::Matrix<Scalar, size, 1>;
         using StateMat = Eigen::Matrix<Scalar, size, size>;
